Time: 25 ms (56.34%) | Memory: 56.5 MB (51.27%) - LeetSync

+++ 2574-count-subarrays-with-median-k/count-subarrays-with-median-k.cpp
@@ -27,7 +27,7 @@ public:
             freq[balance]++;
         }
         balance=0;
-        long long ans=freq[0]+freq[1];
+        int ans=freq[0]+freq[1];
         // left side managing by balancing with right side  by map
         for(int i=pos-1;i>=0;i--)
         {
